gui/requirements.txt
tensorflow
opencv-python
pytesseract
pyzbar
tkinter
requests
scikit-learn
numpy
pandas
mysql-connector-python
